// figma-plugin/src/canvas/types.ts
export type StrokeAlign = 'INSIDE' | 'OUTSIDE' | 'CENTER';
export type TextAlign = 'LEFT' | 'CENTER' | 'RIGHT' | 'JUSTIFIED';
export type TextAutoResize =
  | 'NONE'
  | 'WIDTH_AND_HEIGHT'
  | 'HEIGHT'
  | 'TRUNCATE';
export type FontWeight =
  | 'Thin'
  | 'ExtraLight'
  | 'Light'
  | 'Regular'
  | 'Medium'
  | 'SemiBold'
  | 'Bold'
  | 'ExtraBold'
  | 'Black';
export type ShadowType = 'DROP_SHADOW' | 'INNER_SHADOW';
export type LayoutMode = 'NONE' | 'HORIZONTAL' | 'VERTICAL';
export type AxisSizing = 'FIXED' | 'AUTO';

export interface ShadowDef {
  type?: ShadowType;
  color?: string;
  offsetX?: number;
  offsetY?: number;
  blur?: number;
  spread?: number;
  opacity?: number;
}